refactor: hooks/quries와 apis 분리
