+++ client/src/apis/images.ts
@@ -1,11 +1,11 @@
-import { tilAxios } from '@/apis/utils';
+import client from '@/apis/utils/client';
 
 export interface UploadResponse {
   url: string;
 }
 
 export function upload(imageFile: FormData): Promise<UploadResponse> {
-  return tilAxios
+  return client
     .post('/images/upload', imageFile, {
       headers: { 'content-type': 'multipart/form-data' },
     })
