+++ client/src/apis/index.ts
@@ -1,9 +1,11 @@
 import * as images from './images';
 import * as posts from './posts';
+import * as users from './users';
 
 const apis = {
   images,
   posts,
+  users,
 };
 
 export default apis;
