+++ client/src/apis/posts.ts
@@ -1,5 +1,4 @@
-import { tilAxios } from '@/apis/utils';
-import { getAuthHeader } from '@/apis/utils/auth';
+import client from '@/apis/utils/client';
 import Post from '@/types/Post';
 
 export type CreatePostRequest = Omit<
@@ -8,11 +7,9 @@ export type CreatePostRequest = Omit<
 >;
 
 export function write(param: CreatePostRequest): Promise<number> {
-  return tilAxios
-    .post('/posts', param, getAuthHeader())
-    .then((res) => res.data);
+  return client.post('/posts', param).then((res) => res.data);
 }
 
 export function get(id: number): Promise<Post> {
-  return tilAxios.get(`/posts/${id}`).then((res) => res.data);
+  return client.get(`/posts/${id}`).then((res) => res.data);
 }
