--- client/src/apis/utils/auth.ts
@@ -1,7 +1,0 @@
-import auth from '@/utils/auth';
-
-export const getAuthHeader = (): { headers: { Authorization: string } } => ({
-  headers: {
-    Authorization: `Bearer ${auth.get()}`,
-  },
-});
