--- client/src/apis/utils/index.ts
@@ -1,5 +1,0 @@
-import axios, { AxiosInstance } from 'axios';
-
-export const tilAxios: AxiosInstance = axios.create({
-  baseURL: process.env.API_HOST,
-});
