+++ client/src/hooks/queries/user/useDeleteUser.ts
@@ -5,22 +5,18 @@ import {
   UseMutationResult,
   useQueryClient,
 } from 'react-query';
-import client from '@/utils/api';
 import { AxiosError } from 'axios';
+import apis from '@/apis';
 
 type UseDeleteUserResult = UseMutationResult<void, AxiosError, void> & {
   deleteUser: UseMutateAsyncFunction<void, AxiosError>;
 };
 
-export async function deleteUser(): Promise<void> {
-  return client.delete('/users').then((res) => res.data);
-}
-
 export default function useDeleteUser(
   options?: UseMutationOptions<void, AxiosError>,
 ): UseDeleteUserResult {
   const queryClient = useQueryClient();
-  const mutation = useMutation(deleteUser, {
+  const mutation = useMutation(apis.users.signOut, {
     ...options,
     async onSuccess(data, variables, context) {
       await queryClient.invalidateQueries('users');
