+++ client/src/hooks/queries/user/useLogin.ts
@@ -4,19 +4,10 @@ import {
   UseMutationOptions,
   UseMutationResult,
 } from 'react-query';
-import client from '@/utils/api';
 import { AxiosError } from 'axios';
-import User from '@/types/User';
 import auth from '@/utils/auth';
-
-export interface LoginRequest {
-  email: User['email'];
-  password: string;
-}
-
-export interface LoginResponse {
-  accessToken: string;
-}
+import { LoginRequest, LoginResponse } from '@/apis/users';
+import apis from '@/apis';
 
 type UseLoginResult = UseMutationResult<
   LoginResponse,
@@ -26,15 +17,11 @@ type UseLoginResult = UseMutationResult<
   login: UseMutateAsyncFunction<LoginResponse, AxiosError, LoginRequest>;
 };
 
-export async function login(data: LoginRequest): Promise<LoginResponse> {
-  return client.post('/login', data).then((res) => res.data);
-}
-
 export default function useLogin(
   options?: UseMutationOptions<LoginResponse, AxiosError, LoginRequest>,
 ): UseLoginResult {
   const mutation = useMutation<LoginResponse, AxiosError, LoginRequest>(
-    (data: LoginRequest) => login(data),
+    (data: LoginRequest) => apis.users.login(data),
     {
       ...options,
       onSuccess(data, variables, context) {
