+++ client/src/hooks/queries/user/useUser.ts
@@ -1,22 +1,18 @@
 import { useQuery, UseQueryOptions, UseQueryResult } from 'react-query';
-import client from '@/utils/api';
 import { AxiosError } from 'axios';
 import User from '@/types/User';
 import auth from '@/utils/auth';
+import apis from '@/apis';
 
 export type UseUserResult = UseQueryResult<User, AxiosError> & {
   user?: User;
   loggedOut: boolean;
 };
 
-export async function getUser(): Promise<User> {
-  return client.get<User>('/users').then((res) => res.data);
-}
-
 export default function useUser(
   options?: UseQueryOptions<User, AxiosError>,
 ): UseUserResult {
-  const result = useQuery<User, AxiosError>('users', getUser, {
+  const result = useQuery<User, AxiosError>('users', apis.users.get, {
     enabled: auth.isLoggedIn(),
     ...options,
   });
